Fix build problems; Move filemanager to subdirectory
