+++ src/components/filemanager/index.js
@@ -1,0 +1,2 @@
+export { FileManagerPublisher } from './FileManagerPublisher';
+export FileManager from './FileManager';
